+++ leeger/util/stat_sheet.py
@@ -79,10 +79,14 @@ def leagueStatSheet(league: League, **kwargs) -> AllTimeStatSheet:
     adjustedTeamLuck = SSLAllTimeCalculator.getAdjustedTeamLuck(league, **kwargs)
 
     # check for optional stats
+    totalGames = None
     leagueMedianWins = None
+    opponentLeagueMedianWins = None
     for year in league.years:
         if year.yearSettings.leagueMedianGames is True:
+            totalGames = TeamSummaryAllTimeCalculator.getTotalGames(league, **kwargs)
             leagueMedianWins = GameOutcomeAllTimeCalculator.getLeagueMedianWins(league, **kwargs)
+            opponentLeagueMedianWins = GameOutcomeAllTimeCalculator.getOpponentLeagueMedianWins(league, **kwargs)
             break
 
     return AllTimeStatSheet(gamesPlayed=gamesPlayed, wins=wins, losses=losses, ties=ties, winPercentage=winPercentage,
@@ -95,7 +99,8 @@ def leagueStatSheet(league: League, **kwargs) -> AllTimeStatSheet:
                             opponentScoringShare=opponentScoringShare, maxScore=maxScore, minScore=minScore,
                             scoringStandardDeviation=scoringStandardDeviation, plusMinus=plusMinus,
                             adjustedTeamScore=adjustedTeamScore, adjustedTeamSuccess=adjustedTeamSuccess,
-                            adjustedTeamLuck=adjustedTeamLuck, leagueMedianWins=leagueMedianWins)
+                            adjustedTeamLuck=adjustedTeamLuck, leagueMedianWins=leagueMedianWins, totalGames=totalGames,
+                            opponentLeagueMedianWins=opponentLeagueMedianWins)
 
 
 def yearStatSheet(year: Year, **kwargs) -> YearStatSheet:
@@ -147,9 +152,13 @@ def yearStatSheet(year: Year, **kwargs) -> YearStatSheet:
     teamLuck = SSLYearCalculator.getTeamLuck(year, **kwargs)
 
     # check for optional stats
+    totalGames = None
     leagueMedianWins = None
+    opponentLeagueMedianWins = None
     if year.yearSettings.leagueMedianGames is True:
+        totalGames = TeamSummaryYearCalculator.getTotalGames(year, **kwargs)
         leagueMedianWins = GameOutcomeYearCalculator.getLeagueMedianWins(year, **kwargs)
+        opponentLeagueMedianWins = GameOutcomeYearCalculator.getOpponentLeagueMedianWins(year, **kwargs)
 
     return YearStatSheet(gamesPlayed=gamesPlayed, wins=wins, losses=losses, ties=ties, winPercentage=winPercentage,
                          wal=wal, walPerGame=walPerGame, awal=awal, awalPerGame=awalPerGame, opponentAWAL=opponentAWAL,
@@ -161,4 +170,5 @@ def yearStatSheet(year: Year, **kwargs) -> YearStatSheet:
                          opponentScoringShare=opponentScoringShare, maxScore=maxScore, minScore=minScore,
                          scoringStandardDeviation=scoringStandardDeviation, plusMinus=plusMinus,
                          teamScore=teamScore, teamSuccess=teamSuccess, teamLuck=teamLuck,
-                         leagueMedianWins=leagueMedianWins)
+                         leagueMedianWins=leagueMedianWins, totalGames=totalGames,
+                         opponentLeagueMedianWins=opponentLeagueMedianWins)