+++ test/test_calculator/test_year_calculator/test_AWALYearCalculator.py
@@ -2,6 +2,7 @@ import unittest
 
 from leeger.calculator.year_calculator.AWALYearCalculator import AWALYearCalculator
 from leeger.enum.MatchupType import MatchupType
+from leeger.model.league import YearSettings
 from leeger.model.league.Matchup import Matchup
 from leeger.model.league.Week import Week
 from leeger.model.league.Year import Year
@@ -45,6 +46,35 @@ class TestAWALYearCalculator(unittest.TestCase):
         self.assertEqual(Deci("2.1"), response[teams[3].id])
         self.assertEqual(Deci("2.1"), response[teams[4].id])
         self.assertEqual(Deci("3"), response[teams[5].id])
+
+    def test_getAWAL_leagueMedianGamesIsOn_addsLeagueMedianWinsToAWAL(self):
+        owners, teams = getNDefaultOwnersAndTeams(6)
+
+        matchup1 = Matchup(teamAId=teams[0].id, teamBId=teams[1].id, teamAScore=1, teamBScore=2)
+        matchup2 = Matchup(teamAId=teams[2].id, teamBId=teams[3].id, teamAScore=3, teamBScore=4)
+        matchup3 = Matchup(teamAId=teams[4].id, teamBId=teams[5].id, teamAScore=4, teamBScore=5)
+
+        week1 = Week(weekNumber=1, matchups=[matchup1, matchup2, matchup3])
+
+        matchup1 = Matchup(teamAId=teams[0].id, teamBId=teams[1].id, teamAScore=1, teamBScore=2)
+        matchup2 = Matchup(teamAId=teams[2].id, teamBId=teams[3].id, teamAScore=3, teamBScore=4)
+        matchup3 = Matchup(teamAId=teams[4].id, teamBId=teams[5].id, teamAScore=4, teamBScore=5)
+
+        week2 = Week(weekNumber=2, matchups=[matchup1, matchup2, matchup3])
+
+        matchup1 = Matchup(teamAId=teams[0].id, teamBId=teams[1].id, teamAScore=1, teamBScore=2)
+        matchup2 = Matchup(teamAId=teams[2].id, teamBId=teams[3].id, teamAScore=3, teamBScore=4)
+        matchup3 = Matchup(teamAId=teams[4].id, teamBId=teams[5].id, teamAScore=4, teamBScore=5)
+
+        week3 = Week(weekNumber=3, matchups=[matchup1, matchup2, matchup3])
+
+        yearSettings = YearSettings(leagueMedianGames=True)
+        year = Year(yearNumber=2000, teams=teams, weeks=[week1, week2, week3], yearSettings=yearSettings)
+
+        response = AWALYearCalculator.getAWAL(year)
+
+        self.assertIsInstance(response, dict)
+        self.assertEqual(6, len(response.keys()))
 
     def test_getAWAL_noneIfNoGamesPlayed(self):
         owners, teams = getNDefaultOwnersAndTeams(6)
@@ -480,6 +510,41 @@ class TestAWALYearCalculator(unittest.TestCase):
         self.assertEqual(Deci("0.7"), response[teams[4].id])
         self.assertEqual(Deci("1"), response[teams[5].id])
 
+    def test_getAWALPerGame_leagueMedianGamesIsOn(self):
+        owners, teams = getNDefaultOwnersAndTeams(6)
+
+        matchup1 = Matchup(teamAId=teams[0].id, teamBId=teams[1].id, teamAScore=1, teamBScore=2)
+        matchup2 = Matchup(teamAId=teams[2].id, teamBId=teams[3].id, teamAScore=3, teamBScore=4)
+        matchup3 = Matchup(teamAId=teams[4].id, teamBId=teams[5].id, teamAScore=4, teamBScore=5)
+
+        week1 = Week(weekNumber=1, matchups=[matchup1, matchup2, matchup3])
+
+        matchup1 = Matchup(teamAId=teams[0].id, teamBId=teams[1].id, teamAScore=1, teamBScore=2)
+        matchup2 = Matchup(teamAId=teams[2].id, teamBId=teams[3].id, teamAScore=3, teamBScore=4)
+        matchup3 = Matchup(teamAId=teams[4].id, teamBId=teams[5].id, teamAScore=4, teamBScore=5)
+
+        week2 = Week(weekNumber=2, matchups=[matchup1, matchup2, matchup3])
+
+        matchup1 = Matchup(teamAId=teams[0].id, teamBId=teams[1].id, teamAScore=1, teamBScore=2)
+        matchup2 = Matchup(teamAId=teams[2].id, teamBId=teams[3].id, teamAScore=3, teamBScore=4)
+        matchup3 = Matchup(teamAId=teams[4].id, teamBId=teams[5].id, teamAScore=4, teamBScore=5)
+
+        week3 = Week(weekNumber=3, matchups=[matchup1, matchup2, matchup3])
+
+        yearSettings = YearSettings(leagueMedianGames=True)
+        year = Year(yearNumber=2000, teams=teams, weeks=[week1, week2, week3], yearSettings=yearSettings)
+
+        response = AWALYearCalculator.getAWALPerGame(year)
+
+        self.assertIsInstance(response, dict)
+        self.assertEqual(6, len(response.keys()))
+        self.assertEqual(Deci("0"), response[teams[0].id])
+        self.assertEqual(Deci("0.1"), response[teams[1].id])
+        self.assertEqual(Deci("0.2"), response[teams[2].id])
+        self.assertEqual(Deci("0.85"), response[teams[3].id])
+        self.assertEqual(Deci("0.85"), response[teams[4].id])
+        self.assertEqual(Deci("1"), response[teams[5].id])
+
     def test_getAWALPerGame_noneIfNoGamesPlayed(self):
         owners, teams = getNDefaultOwnersAndTeams(6)
 
@@ -822,6 +887,41 @@ class TestAWALYearCalculator(unittest.TestCase):
         self.assertEqual(Deci("3"), response[teams[4].id])
         self.assertEqual(Deci("2.1"), response[teams[5].id])
 
+    def test_getOpponentAWAL_leagueMedianGamesIsOn_countsLeagueMedianGames(self):
+        owners, teams = getNDefaultOwnersAndTeams(6)
+
+        matchup1 = Matchup(teamAId=teams[0].id, teamBId=teams[1].id, teamAScore=1, teamBScore=2)
+        matchup2 = Matchup(teamAId=teams[2].id, teamBId=teams[3].id, teamAScore=3, teamBScore=4)
+        matchup3 = Matchup(teamAId=teams[4].id, teamBId=teams[5].id, teamAScore=4, teamBScore=5)
+
+        week1 = Week(weekNumber=1, matchups=[matchup1, matchup2, matchup3])
+
+        matchup1 = Matchup(teamAId=teams[0].id, teamBId=teams[1].id, teamAScore=1, teamBScore=2)
+        matchup2 = Matchup(teamAId=teams[2].id, teamBId=teams[3].id, teamAScore=3, teamBScore=4)
+        matchup3 = Matchup(teamAId=teams[4].id, teamBId=teams[5].id, teamAScore=4, teamBScore=5)
+
+        week2 = Week(weekNumber=2, matchups=[matchup1, matchup2, matchup3])
+
+        matchup1 = Matchup(teamAId=teams[0].id, teamBId=teams[1].id, teamAScore=1, teamBScore=2)
+        matchup2 = Matchup(teamAId=teams[2].id, teamBId=teams[3].id, teamAScore=3, teamBScore=4)
+        matchup3 = Matchup(teamAId=teams[4].id, teamBId=teams[5].id, teamAScore=4, teamBScore=5)
+
+        week3 = Week(weekNumber=3, matchups=[matchup1, matchup2, matchup3])
+
+        yearSettings = YearSettings(leagueMedianGames=True)
+        year = Year(yearNumber=2000, teams=teams, weeks=[week1, week2, week3], yearSettings=yearSettings)
+
+        response = AWALYearCalculator.getOpponentAWAL(year)
+
+        self.assertIsInstance(response, dict)
+        self.assertEqual(6, len(response.keys()))
+        self.assertEqual(Deci("0.6"), response[teams[0].id])
+        self.assertEqual(Deci("0"), response[teams[1].id])
+        self.assertEqual(Deci("5.1"), response[teams[2].id])
+        self.assertEqual(Deci("1.2"), response[teams[3].id])
+        self.assertEqual(Deci("6"), response[teams[4].id])
+        self.assertEqual(Deci("5.1"), response[teams[5].id])
+
     def test_getOpponentAWAL_noneIfNoGamesPlayed(self):
         owners, teams = getNDefaultOwnersAndTeams(6)
 
@@ -1101,7 +1201,8 @@ class TestAWALYearCalculator(unittest.TestCase):
 
         week1 = Week(weekNumber=1, matchups=[matchup1, matchup2, matchup3])
 
-        year = Year(yearNumber=2000, teams=[teams[0], teams[1], teams[2], teams[3], teams[4], teams[5]], weeks=[week1])
+        year = Year(yearNumber=2000, teams=[teams[0], teams[1], teams[2], teams[3], teams[4], teams[5]],
+                    weeks=[week1])
 
         response = AWALYearCalculator.getOpponentAWAL(year)
 
@@ -1123,7 +1224,8 @@ class TestAWALYearCalculator(unittest.TestCase):
 
         week1 = Week(weekNumber=1, matchups=[matchup1, matchup2, matchup3])
 
-        year = Year(yearNumber=2000, teams=[teams[0], teams[1], teams[2], teams[3], teams[4], teams[5]], weeks=[week1])
+        year = Year(yearNumber=2000, teams=[teams[0], teams[1], teams[2], teams[3], teams[4], teams[5]],
+                    weeks=[week1])
 
         response = AWALYearCalculator.getOpponentAWAL(year)
 
@@ -1145,7 +1247,8 @@ class TestAWALYearCalculator(unittest.TestCase):
 
         week1 = Week(weekNumber=1, matchups=[matchup1, matchup2, matchup3])
 
-        year = Year(yearNumber=2000, teams=[teams[0], teams[1], teams[2], teams[3], teams[4], teams[5]], weeks=[week1])
+        year = Year(yearNumber=2000, teams=[teams[0], teams[1], teams[2], teams[3], teams[4], teams[5]],
+                    weeks=[week1])
 
         response = AWALYearCalculator.getOpponentAWAL(year)
 
@@ -1167,7 +1270,8 @@ class TestAWALYearCalculator(unittest.TestCase):
 
         week1 = Week(weekNumber=1, matchups=[matchup1, matchup2, matchup3])
 
-        year = Year(yearNumber=2000, teams=[teams[0], teams[1], teams[2], teams[3], teams[4], teams[5]], weeks=[week1])
+        year = Year(yearNumber=2000, teams=[teams[0], teams[1], teams[2], teams[3], teams[4], teams[5]],
+                    weeks=[week1])
 
         response = AWALYearCalculator.getOpponentAWAL(year)
 
@@ -1255,6 +1359,41 @@ class TestAWALYearCalculator(unittest.TestCase):
         self.assertEqual(Deci("0.4"), response[teams[3].id])
         self.assertEqual(Deci("1"), response[teams[4].id])
         self.assertEqual(Deci("0.7"), response[teams[5].id])
+
+    def test_getOpponentAWALPerGame_leagueMedianGamesIsOn(self):
+        owners, teams = getNDefaultOwnersAndTeams(6)
+
+        matchup1 = Matchup(teamAId=teams[0].id, teamBId=teams[1].id, teamAScore=1, teamBScore=2)
+        matchup2 = Matchup(teamAId=teams[2].id, teamBId=teams[3].id, teamAScore=3, teamBScore=4)
+        matchup3 = Matchup(teamAId=teams[4].id, teamBId=teams[5].id, teamAScore=4, teamBScore=5)
+
+        week1 = Week(weekNumber=1, matchups=[matchup1, matchup2, matchup3])
+
+        matchup1 = Matchup(teamAId=teams[0].id, teamBId=teams[1].id, teamAScore=1, teamBScore=2)
+        matchup2 = Matchup(teamAId=teams[2].id, teamBId=teams[3].id, teamAScore=3, teamBScore=4)
+        matchup3 = Matchup(teamAId=teams[4].id, teamBId=teams[5].id, teamAScore=4, teamBScore=5)
+
+        week2 = Week(weekNumber=2, matchups=[matchup1, matchup2, matchup3])
+
+        matchup1 = Matchup(teamAId=teams[0].id, teamBId=teams[1].id, teamAScore=1, teamBScore=2)
+        matchup2 = Matchup(teamAId=teams[2].id, teamBId=teams[3].id, teamAScore=3, teamBScore=4)
+        matchup3 = Matchup(teamAId=teams[4].id, teamBId=teams[5].id, teamAScore=4, teamBScore=5)
+
+        week3 = Week(weekNumber=3, matchups=[matchup1, matchup2, matchup3])
+
+        yearSettings = YearSettings(leagueMedianGames=True)
+        year = Year(yearNumber=2000, teams=teams, weeks=[week1, week2, week3], yearSettings=yearSettings)
+
+        response = AWALYearCalculator.getOpponentAWALPerGame(year)
+
+        self.assertIsInstance(response, dict)
+        self.assertEqual(6, len(response.keys()))
+        self.assertEqual(Deci("0.1"), response[teams[0].id])
+        self.assertEqual(Deci("0"), response[teams[1].id])
+        self.assertEqual(Deci("0.85"), response[teams[2].id])
+        self.assertEqual(Deci("0.2"), response[teams[3].id])
+        self.assertEqual(Deci("1"), response[teams[4].id])
+        self.assertEqual(Deci("0.85"), response[teams[5].id])
 
     def test_getOpponentAWALPerGame_noneIfNoGamesPlayed(self):
         owners, teams = getNDefaultOwnersAndTeams(6)
@@ -1562,4 +1701,3 @@ class TestAWALYearCalculator(unittest.TestCase):
         self.assertEqual(Deci("0.4"), response[teams[3].id])
         self.assertEqual(Deci("1"), response[teams[4].id])
         self.assertEqual(Deci("0.85"), response[teams[5].id])
-